Completed parsing of all Post class.
Post class objects are created and work for automated posts. Still want to sort the objects in picture. Need to find an affective way of parsing the on post text, this will be developed alongside the search algorithm. Need to organize code and make a separate JS file for all the 'post parsing' related code.